+++ src/signal/seek.rs
@@ -19,7 +19,7 @@ use strum::{
 /// # Size
 /// ```rust
 /// # use sansan::signal::*;
-/// assert_eq!(std::mem::size_of::<Seek>(), 16);
+/// assert_eq!(std::mem::size_of::<Seek>(), 8);
 /// ```
 #[derive(Copy,Clone,Debug,PartialEq,PartialOrd)]
 #[derive(AsRefStr,Display,EnumCount,EnumIter,EnumString,EnumVariantNames,EnumDiscriminants,IntoStaticStr)]
@@ -29,16 +29,16 @@ use strum::{
 #[serde(rename_all = "snake_case")]
 pub enum Seek {
 	/// Seek to an absolute second timestamp
-	Absolute(f64),
+	Absolute(f32),
 	/// Seek forwards a specified amount of seconds
-	Forward(f64),
+	Forward(f32),
 	/// Seek backwards a specified amount of seconds
-	Backward(f64),
+	Backward(f32),
 }
 
 /// The (second) timestamp `Decode` successfully
 /// set the time to after a seek operation.
-pub(crate) type SeekedTime = f64;
+pub(crate) type SeekedTime = f32;
 
 //---------------------------------------------------------------------------------------------------- SeekError
 /// TODO
